moving image into its own widget
