+++ packages/blog_ui/lib/src/widgets/widgets.dart
@@ -1,3 +1,4 @@
 export 'author_avatar.dart';
 export 'author_tile.dart';
 export 'blog_card.dart';
+export 'featured_image.dart';
